接口文档上传

+++ meituan_server/src/main/java/com/example/meituan/domain/Food.java
@@ -56,7 +56,7 @@ public class Food implements Serializable {
      * 平均评分
      */
     @TableField("avgScore")
-    private Integer avgScore;
+    private long avgScore;
 
     /**
      * 总评论数
